Split Footprint submodules into own files
