+++ src/faebryk/library/DIP.py
@@ -2,8 +2,9 @@
 # SPDX-License-Identifier: MIT
 
 from faebryk.library.can_attach_via_pinmap_equal import can_attach_via_pinmap_equal
-from faebryk.library.Footprint import Footprint, Pad
+from faebryk.library.Footprint import Footprint
 from faebryk.library.has_equal_pins_in_ifs import has_equal_pins_in_ifs
+from faebryk.library.Pad import Pad
 from faebryk.libs.util import times
