+++ src/faebryk/library/has_linked_pad.py
@@ -3,10 +3,12 @@
 
 from abc import abstractmethod
 
-from faebryk.library.FootprintTrait import FootprintTrait
+from faebryk.core.core import (
+    ModuleInterfaceTrait,
+)
 from faebryk.library.Pad import Pad
 
 
-class has_equal_pins(FootprintTrait):
+class has_linked_pad(ModuleInterfaceTrait):
     @abstractmethod
-    def get_pin_map(self) -> dict[Pad, str]: ...
+    def get_pad(self) -> Pad: ...
